+++ reco/detectors/trd/unpack/CbmTrdUnpackAlgoFasp2D.cxx
@@ -146,7 +146,8 @@ bool CbmTrdUnpackAlgoFasp2D::unpack(const fles::Timeslice* ts, std::uint16_t ico
   //Double_t fdMsSizeInNs = 1.28e6;
 
   auto msdesc = ts->descriptor(icomp, imslice);
-  if (VERBOSE) printf("time start %lu\n", msdesc.idx);
+  // Cast required to silence a warning on macos (there a uint64_t is a llu)
+  if (VERBOSE) printf("time start %lu\n", static_cast<size_t>(msdesc.idx));
   // define time wrt start of time slice in TRD/FASP clks [80 MHz]
   fTime[0] = ULong64_t((msdesc.idx - fTsStartTime) / 12.5);
